requirements.txt
gmqtt==0.6.2
PyYAML==5.3.1
voluptuous==0.11.7
